fix saveSystem being inside callbacks
